+++ src/components/ProgressBar/interfaces.ts
@@ -1,0 +1,4 @@
+export interface IProgressBar {
+  backgroundColor: string;
+  completed: number;
+}