+++ src/index.tsx
@@ -35,6 +35,8 @@ import Breadcrumbs from './components/Breadcrumbs';
 import Orderer from './components/Orderer';
 import Infotip from './components/Infotip';
 import DatePicker from './components/DatePicker';
+import FileUpload from './components/FileUpload';
+import ProgressBar from './components/ProgressBar';
 import Select from './components/Select';
 import LabelSwitch from './components/LabelSwitch';
 
@@ -76,6 +78,8 @@ export {
   Orderer,
   Infotip,
   DatePicker,
+  FileUpload,
+  ProgressBar,
   Select,
   LabelSwitch,
 };