+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.32.0] - 2021-08-30
+
+- Added FileUpload component
+- Added ProgressBar component
+
 ## [1.31.0] - 2021-08-21
 
 - Added Select component